Fixed: image complementary attributes like IMAGEPRESS, IMAGEHIGHLIGHT, IMAGEINACTIVE, and others, in IupGLControls.
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@2483 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/srcglcontrols/iup_glfont.c
@@ -584,7 +584,7 @@ static int iGLFontGetFontAveWidth(FTGLfont* font)
     iGLFontConvertToUTF8(sample, 256 - 32);
     if (utf8mode) IupSetGlobal("UTF8MODE", "Yes");
 
-    i = strlen(utf8_buffer);
+    i = (int)strlen(utf8_buffer);
     memcpy(sample, utf8_buffer, i + 1);
     first = 0;
   }
